client/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
pika
psycopg2-binary   # si tu utilises PostgreSQL
gunicorn
